test(panelbar): remove and update panelbar tests
