+++ packages/html/src/panelbar/tests/panelbar-rtl.html
@@ -8,6 +8,6 @@
     <script src="/packages/html/assets/scripts.js"></script>
 </head>
 <body id="app" class="k-body">
-    <script src="/packages/html/dist/panelbar/tests/panelbar-universal.js"></script>
+    <script src="/packages/html/dist/panelbar/tests/panelbar-rtl.js"></script>
 </body>
 </html>
