+++ packages/html/src/panelbar/tests/panelbar.html
@@ -1,13 +1,13 @@
 <!DOCTYPE html>
 <html lang="en" class="k-no-animations">
 <head>
-    <title>PanelBar React</title>
+    <title>PanelBar Universal</title>
     <meta charset="utf-8" />
     <link rel="stylesheet" data-role="kendo-theme" href="/packages/default/dist/all.css"/>
     <link rel="stylesheet" href="/packages/html/assets/styles.css"/>
     <script src="/packages/html/assets/scripts.js"></script>
 </head>
 <body id="app" class="k-body">
-    <script src="/packages/html/dist/panelbar/tests/panelbar-react.js"></script>
+    <script src="/packages/html/dist/panelbar/tests/panelbar.js"></script>
 </body>
 </html>
